requirements.txt
sqlalchemy==1.4.46
numpy==1.24.3
python-dateutil==2.8.2
panel==1.2.1
holoviews==1.17.1
param==1.13.0
pandas==2.0.3
speedtest-cli==2.1.3
bokeh==3.2.2

